Completely reworked the code and moved game logic into game objects

+++ SpaceInvaders/SpaceInvaders/InvadersConstants.hpp
@@ -38,7 +38,7 @@ static constexpr int SHOOTING_ENEMY_MAX_HORIZONTAL_SPEED = 100; // horizontal mo
 static constexpr int SHOOTING_ENEMY_MAX_HORIZONTAL_RANGE = 500; // how many pixels to the left and right from original position can the enemy move
 static constexpr int SHOOTING_ENEMY_MIN_HORIZONTAL_RANGE = 200; // how many pixels to the left and right from original position can the enemy move
 static constexpr int SHOOTING_ENEMY_MAX_Y_POSITION = 300; // distance of the ShootingEnemy from the top of screen
-static constexpr int SHOOTING_ENEMY_MIN_Y_POSITION = ENEMY_SIZE / 2 + 10; // distance of the ShootingEnemy from the top of screen
+static constexpr int SHOOTING_ENEMY_MIN_Y_POSITION = (int)(ENEMY_SIZE / 2) + 10; // distance of the ShootingEnemy from the top of screen
 static constexpr int SHOOTING_ENEMY_MAX_ROCKET_TIME = 10; // time between shooting rockets (seconds)
 static constexpr int SHOOTING_ENEMY_MIN_ROCKET_TIME = 1; // time between shooting rockets (seconds)
